Fix C4703 uninitialized ctxt and remaining 64-bit conversion warnings
- Initialized ctxt = NULL in frontend_encrypt and added error handling for unknown file types.
- Added explicit (u32) casts to ftell and header_len assignments to resolve conversion warnings in np.cpp, sce.cpp, and util.cpp.
- Ensured overall compatibility with MSVC code generation.

+++ np.cpp
@@ -221,7 +221,7 @@ BOOL np_sign_file(s8 *fname)
 		return FALSE;
 
 	fseek(fp, 0, SEEK_END);
-	length = ftell(fp);
+	length = (u32)ftell(fp);
 
 	padding = length % 0x10;
 	if(padding > 0)
